requirements.txt
flwr==1.7.0
torch>=2.0.0
torchvision
numpy
pyyaml
matplotlib